backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
anthropic>=0.28.0
python-dotenv>=1.0.1
pydantic>=2.7.0
aiohttp>=3.9.5
redis>=5.0.4
celery>=5.4.0
fpdf2>=2.7.9
python-pptx>=0.6.23
pytest>=8.0.0
pytest-mock>=3.12.0
httpx>=0.27.0
